+++ src/components/index.js
@@ -1,7 +1,6 @@
 export { default as FilterNavigation } from './FilterNavigation';
 export { default as RepeatableField } from './RepeatableField';
 export { default as InstanceFilters, instanceFilterRenderer } from './InstanceFilters';
-export { default as InstanceFiltersBrowse, instanceFilterBrowseRenderer } from './InstanceBrowseFilters';
 export { default as HoldingsRecordFilters, holdingsRecordFilterRenderer } from './HoldingsRecordFilters';
 export { default as ItemFilters, itemFilterRenderer } from './ItemFilters';
 export { default as InstancesList } from './InstancesList';
@@ -18,3 +17,5 @@ export { default as WarningMessage } from './WarningMessage';
 export { default as CalloutRenderer } from './CalloutRenderer';
 export { default as ViewSource } from './ViewSource';
 export { default as AdministrativeNoteList } from './AdministrativeNoteList';
+export { default as MultiSelectionFacet } from './MultiSelectionFacet';
+export { default as FacetOptionFormatter } from './FacetOptionFormatter';